requirements.txt
discord.py
pytesseract
opencv-python-headless
deepface
Pillow
requests
